Make `ProjectileCS`s stop when hitting shield
This is based on jetpack-maniac's fork/patch from A17, but manually applied to the latest B18 version.  It seems Combat Extended does something to make its projectiles not count as instances of Projectile, which makes the shield generators ignore them.  There's probably a more elegant solution to handle two possible types that expose essentially the same interface, but I'll leave that to someone with a C# IDE who actually knows C# to implement better.